Add library image for notary-signer
This is based on swarm-library-image.

Unfortunately, the notary-server binary needs to be checked in -
otherwise "docker build" wouldn't work without preparation, as is
required of official images. The binary checked in here is from notary
a95349457ab2a1f36209596bebe9464dc719a335. It will need to be updated to
the release version after each release (using the update.sh script).

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>